genai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.3.25
langchain-core==0.3.65
langchain-community==0.3.25
langchain-huggingface==0.3.0
PyMuPDF==1.23.7
pypdf==3.17.1
faiss-cpu==1.7.4
sentence-transformers==2.3.1 # Upgraded from 2.2.2
python-dotenv==1.0.0
numpy<2.0
torch==2.1.1
torchvision==0.16.1
requests==2.31.0
langsmith==0.3.45